requirements.txt
aiohttp==3.7.4
baidu_aip==4.16.11
Brotli==1.0.9
keyboard==0.13.5
numpy==1.25.1
numpy==1.22.0
openai==0.27.8
pure_protobuf==2.3.0
PyAudio==0.2.13
pygtrans==1.5.3
pygtrans==1.5.2
Requests==2.31.0
snownlp==0.12.3
SpeechRecognition==3.10.0
sympy==1.12
sympy==1.11.1
typing_extensions==4.7.1
websockets==11.0.3
